delete legacy folder: - remove legacy things used in FullGrid - move basis functions to hierarchization - create PowerOfTwo.hpp
