+++ distributedcombigrid/src/sgpp/distributedcombigrid/hierarchization/CombiBasisFunctionBasis.hpp
@@ -2,7 +2,6 @@
 #define COMBIBASISFUNCTION_HPP_
 
 #include <boost/serialization/access.hpp>
-#include <sgpp/distributedcombigrid/legacy/combigrid_utils.hpp>
 
 namespace combigrid {
 
@@ -20,18 +19,6 @@ class BasisFunctionBasis {
   BasisFunctionBasis() = default;
   /** virtual Dtor */
   virtual ~BasisFunctionBasis() = default;
-
-  /** first method which returns the contribution of the first point in the 1D
-   * cell
-   * @param coord  1D coordonate idealy should be [0,1] but for extrapolation
-   * could be different [-1,2]*/
-  virtual double functionEval1(double coord) const = 0;
-
-  /** second method which returns the contribution of the second point in the 1D
-   * cell
-   * @param coord  1D coordonate idealy should be [0,1] but for extrapolation
-   * could be different [-1,2]*/
-  virtual double functionEval2(double coord) const = 0;
 
  private:
   friend class boost::serialization::access;
