+++ distributedcombigrid/src/sgpp/distributedcombigrid/hierarchization/CombiLinearBasisFunction.hpp
@@ -2,8 +2,7 @@
 #define COMBILINEARBASISFUNCTION_HPP_
 
 #include <boost/serialization/base_object.hpp>
-#include <sgpp/distributedcombigrid/legacy/CombiBasisFunctionBasis.hpp>
-#include <sgpp/distributedcombigrid/legacy/combigrid_utils.hpp>
+#include <sgpp/distributedcombigrid/hierarchization/CombiBasisFunctionBasis.hpp>
 
 namespace combigrid {
 
@@ -13,18 +12,6 @@ class LinearBasisFunction : public combigrid::BasisFunctionBasis {
   /** empty Ctror */
   LinearBasisFunction() = default;
   virtual ~LinearBasisFunction() = default;
-
-  /** first method which returns the contribution of the first point in the 1D
-   * cell
-   * @param coord  1D coordonate idealy should be [0,1] but for extrapolation
-   * could be different [-1,2]*/
-  double functionEval1(double coord) const override { return (1.0 - coord); }
-
-  /** second method which returns the contribution of the second point in the 1D
-   * cell
-   * @param coord  1D coordonate idealy should be [0,1] but for extrapolation
-   * could be different [-1,2]*/
-  double functionEval2(double coord) const override { return (coord); }
 
   // /** return the default basis function*/
   // static const BasisFunctionBasis* getDefaultBasis() { return defaultBasis_; }
