+++ distributedcombigrid/src/sgpp/distributedcombigrid/manager/CombiParameters.hpp
@@ -2,10 +2,11 @@
 #define SRC_SGPP_COMBIGRID_MANAGER_COMBIPARAMETERS_HPP_
 
 #include <boost/serialization/map.hpp>
-#include "sgpp/distributedcombigrid/legacy/CombiLinearBasisFunction.hpp"
+#include "sgpp/distributedcombigrid/hierarchization/CombiLinearBasisFunction.hpp"
 #include "sgpp/distributedcombigrid/mpi/MPISystem.hpp"
 #include "sgpp/distributedcombigrid/utils/LevelSetUtils.hpp"
 #include "sgpp/distributedcombigrid/utils/LevelVector.hpp"
+#include "sgpp/distributedcombigrid/utils/PowerOfTwo.hpp"
 #include "sgpp/distributedcombigrid/utils/Types.hpp"
 namespace combigrid {
