+++ distributedcombigrid/src/sgpp/distributedcombigrid/utils/BoostExports.hpp
@@ -5,7 +5,7 @@
 #include "sgpp/distributedcombigrid/fault_tolerance/FaultCriterion.hpp"
 #include "sgpp/distributedcombigrid/fault_tolerance/StaticFaults.hpp"
 #include "sgpp/distributedcombigrid/fault_tolerance/WeibullFaults.hpp"
-#include "sgpp/distributedcombigrid/legacy/CombiLinearBasisFunction.hpp"
+#include "sgpp/distributedcombigrid/hierarchization/CombiLinearBasisFunction.hpp"
 
 // this header should be included once for every compilation unit; if there are
 // "not registered" or "not exported"-type errors, maybe this header was called before all
