+++ distributedcombigrid/src/sgpp/distributedcombigrid/utils/LevelSetUtils.hpp
@@ -3,22 +3,10 @@
 #include <vector>
 
 #include "sgpp/distributedcombigrid/utils/LevelVector.hpp"
+#include "sgpp/distributedcombigrid/utils/PowerOfTwo.hpp"
 #include "sgpp/distributedcombigrid/utils/Types.hpp"
 
 namespace combigrid {
-
-// compute powers of two quickly by bit-shifting
-inline IndexType powerOfTwoByBitshift(LevelType x) {
-  assert(static_cast<long unsigned int>(x) < sizeof(IndexType) * 8);
-  if (x == 0) {
-    return 1;
-  } else if (x > 0) {
-    return 2 << (x - 1);
-  } else if (x < 0) {
-    throw std::runtime_error("powerOfTwoByBitshift: negative argument");
-  }
-  return -1;
-}
 
 inline IndexType getNumDofNodal(const LevelType& l_i, const BoundaryType& boundary) {
   assert(l_i > -1);
